v9, drbdadm adjust: relaxed parser
Currently we might have this situation:
- User sets an option with drbdsetup (e.g., c-max-rate) to a value that
is actually not in the valid range for this option. This is fine and
should be allowed.
- User sets a valid value in the configuration and tries to "drbdadm
adjust". This fails because adjust reads back the output of drbdsetup
and the parser fails because the value parsed is not in a valid range
and the adjust process was aborted.

Two possible fixes:
- Do not allow drbdsetup to even set this value.
- Relax "drbdadm adjust" parsing.

For "relaxed parsing" there is already "config_valid = 2", make use of
it in case an error is detected.

In general we follow a "parse and report every error" approach, therfore
do not exit() in this early stage, just set config_valid to 0.