v9, drbdadm adjust: filter transport option for drbdsetup
If the user set "transport" as net-option and then changed another
net-option in the net section, "drbdadm adjust" called drbdsetup with
the "--transport" parameter even if the transport obviously was not
changed. In the kernel part the transport is fixed, resending the
parameter is a bad idea. Resending is now eliminated by filtering the
transport option in the call path where it does not make sense.