drbdadm: Introduce a data structure for network-paths
Creating a single implicit path for each connection.

Strictly speaking each path could have different proxy settings.
This is ignored right now. I simply assumed that in case proxy
is involved only a single path exists.

The adjust logic for multiple pathes is rather incomplete with
this commit. Right now it will simply re-create all pathes in
case an address mismatch is found.
This is incomplete and needs to be implemented correctly later.