drbdmeta: drop usage of strndupa() which is not provided by musl
The musl C library does not support strndupa() which is an
implementation of strndup() that allocates its buffer on the stack.

This patch changes drbd_str_disk() to compute the string length up to
the first slash (if any) and use the length to match the string with
the state names.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>
Signed-off-by: Roland Kammerer <roland.kammerer@linbit.com>