build: fix building with/without manual pages.
This fixes a regression introduced with a previous patch that should
allow building without man pages. The problem was that it required an
additional step "make doc". This additional step was for example not
present in the Debian build environment.

This allows the usual "./configure && make && make install" process
without requiring "make doc". Building without man pages is handled by
"./configure --without-manual".