build: fix BITS_PER_LONG define again
When changing stuff for "compat" reasons, do not just change the set of
supported systems, but make sure the new set is a superset of what had
been supported before.

Some older glibc apparently do not have __SIZEOF_LONG__
but they do have __WORDSIZE.

Only use __SIZEOF_LONG__ where it is available.
Use __WORDSIZE where it is not.

(I don't want to have autoconf figure it out for me, but if you really
need to, there is a SIZEOF_LONG_INT test for autoconf)