drbd.ocf: new parameter unfence_if_all_uptodate
Allow a monitor action (or otherwise triggered "update_master_score")
to call an unfence hook, if it finds itself Primary and all local and
remote volumes show as UpToDate.

This is useful together with the crm-fence-peer.sh script,
to remove a stale fencing constraint that happens to still be around
for whatever reason. It may also be useful for other resource-level
fencing methods.