drbdadm,v9: let command line overrule config
Currently, drbdadm generates strange drbdsetup options if an option is
set in the config and on the commandline.

Example: protocol set to 'C' in res file and command line is "drbdadm
net-options foo --protocol=A". In this cases "--protocol" gets added
twice.

This removes duplicates and lets the option specified on the command
line win.