drbdmeta: introduce MDF_PEER_DEVICE_SEEN
A flag in meta data to persist if we have ever seen a certain
peer node with a backing device attached.

If we did not see a backing device, ever, we assume that it
is diskless on purpose (a "DRBD client"), and can ignore it
while calculating the maximum agreeable device size.