debian: add udev as build dependency
All our chroots have udev installed, so this should not tackle us. This
change is mainly done for the PPA and that the debian directories don't
diverge. Without it, we generate PPA packages that contain disabled udev
rules. This was necessary due to a very recent change in the PPA build
environment.