init script: empty default-start, 0 1 6 default-stop
And revert the "update-rc.d disable" to "defaults" in debian/rules.
If someone has reason to use the init script, let him do so,
but do not by default enable it.
In the typical use case, some kind of cluster management is responsible
for bringing up/shutting down the DRBD resources.