new peer_device option --bitmap=no
With that option tell the kernel to not allocate a bitmap for that
device for that peer. Drbdadm uses this option in case the peer's
device is marked a permanently diskless in the configuration file.

In order that this change is effective drbdadm needs to
create minors first, then create peers and set peer-device-options
before attaching or connecting.