build: Integrate drbdmon in build environment
Drbdmon is enabled by default. As the tool requires a modern C++11
compiler, we check for compatibility at ./configure time and disable
building drbdmon if we don't find a compatible compiler.

As we are building for old architectures, this introduces two macros
(in m4/). This avoids bumping the minimal autoconf version.
Additionally, not even recent autotool get it right (Debian Stretch), so
keep them locally.

Signed-off-by: Roland Kammerer <roland.kammerer@linbit.com>