Ensure effective/saved user/group ids == real user/group ids
drbdsetup is now started from ${PATH} instead of a
hardcoded secure path.

As the ${PATH} variable is under the control of the user running
the drbdmon utility, this change is supposed to prevent a local
elevation of privilege exploit in the case that someone has
installed the drbdmon utility with the suid or sgid bit set on
the executable.

The effective and saved user ids are adjusted to match the
real user id and the effective and saved group ids are adjusted
to match the real group id.
If adjusting the ids fails, the program prints an error message
and exits.