Fixed potential buffer index underflow
url_encode() performed a right shift on a signed char, and the
behavior of doing that is implementation defined - it might
perform an arithmetic or logical shift.
(For example, gcc 4.7.2 on Linux performs an arithmetic shift)

For any character with a value equal to or greater than 0x80
(which represents a negative signed char value), the result
of the shift operation is a negative value too, if the compiler
uses an arithmetic shift.

Fixed by changing the type of the value being shifted to
unsigned char.