doc, drbdmon, v9: Fix symlink generation
We have the following situation:
- Every utils version generates its own documentation and mangles it's
  version into the path (e.g., drbdmeta-8.4.8, drbdmeta, version 8.4,
  section 8).

- Then we create symlinks to the according version of the man page (e.g,
  drbdmeta.8 -> drbdmeta-8.4.8). As there are still more drbd8 users, we
  link to the version 8 man pages by default.

This works perfectly fine for man pages that exist in different
versions, but creates a dead symlink for tools/man pages that only exist
in v9 like drbdmon.

This fixes symlink creation and while at it introduces a new variable
that replaces hardcoded "9.0" strings, which might be nice if the
Makefile.in is copied over/changed to a new version in the future.