postparse: Fix NULL ptr deref in fixup_peer_devices()
So far commands like "drbdadm adjust -d" triggered a segfault if a
resource had a connection, but no paths. This can manually be triggered
by something like:

$ drbdadm up .drbdctrl
$ drbdadm disconnect .drbdctrl
$ drbdsetup del-path .drbdctrl 1 10.43.70.115:6999 10.43.70.116:6999
$ drbdadm adjust -d .drbdctrl
segmentation fault

This can be fixed by checking if a connection actually has a path, which
then results in the following:

$ drbdadm adjust -d .drbdctrl
$ drbdsetup new-path .drbdctrl 1 ipv4:10.43.70.115:6999 ipv4:10.43.70.116:6999
$ drbdsetup peer-device-options .drbdctrl 1 0 --set-defaults
$ drbdsetup peer-device-options .drbdctrl 1 1 --set-defaults
$ [...]
$ drbdsetup connect .drbdctrl 1

Thanks to Robert for spotting it and preparing a first version.