Enable root to drop privileges using setuid/setgid
Lars pointed out that someone running the program
as root might attempt to drop privileges by using
suid/sgid on the drbdmon executable. Although this
is not something that can normally be expected to
work securely, this change adds support for using
the suid/sgid mechanism for that purpose.

If the program is executed by real user root, but the
effective user id does not match the real user id, set
the real user id to the effective user id, and ensure
that the saved user id is set accordingly too.
Perform an analogous action for group id adjustment.