parser, v9: hash table -> binary tree
This switches the hash table used for the parser to a binary tree. The
main advantage is that the tree just grows, while the hash table had a
fixed size. To my surprise this is faster and consumes less memory than
the hash table. Tested on power8 machines with some thousand resources.

This also moves two functions from shared parser code to their respective
version subdirectory as this is a v9 only change. For v84 the hash table
is good enough.