build: work around broken autoconf on RHEL6 alikes
On RHEL6 the default autoconf version is 2.63, which is broken because
it does not include macros specified via AC_CONFIG_MACRO_DIR. Installing
"autoconf213" was a workaround for that. Newer autoconf versions
(RHEL7/Debian/...) are fine with AC_CONFIG_MACRO_DIR, which is the
actual way to go, even if it has a "funny"[1] history.

[1] https://lists.gnu.org/archive/html/autoconf/2013-08/msg00013.html