drbdadm, v9: make sure real_path can be freed
Under certain conditions we inserted a pointer to a value that was not
solely under the control of the binary tree. Now we make sure that the
key/value (real_path) is always malloced and therefore under our
control.

Thanks to Robert for spotting it and preparing a first version.