drbdmeta: fix dump-md bitmap RLE and runaway lines
If the buffer wrap on huge bitmaps happend at an "unfortunate" offset
wrt. number of bitmap slots (max_peers) and half-64bit words,
all the modulo logic on the next iteration was off, no more
run-length encoding happens, and even output of newlines was
skipped, resulting in VERY long lines and huge dumps.

The easy fix is to align the max_chunk_size per iteration on a multiple
of number of bitmap slots (max_peers) and 8 Bytes (because of 64bit
dump output of the 32bit interleaved bitmap words per peer).
Actually, align to max_peers * 4k (because of direct IO).