drbdsetup,v9,json: fix JSON output for single res
So far if "drbdsetup status --json" was called with a single resource
and there were further resources in the internal resource list, an
additional ',' was printed. That happened because "resource->next" was
!NULL, so we printed the ',', but then called "continue" in the for-loop
without printing output.

A first version fixed that by doing the same early-continue check for
resource->next before printing the ','. This introduces either code
duplication or a nasty macro.

This version fixes it by changing the logic. We print the ',' before we
print JSON output, but skip the first one.