drbdmeta: When creating metadata set bitmap to all zero
Before the patch it was left uninitialized. Initializing it to all zero
helps for the (uncommon) use case that one uses set-gi to avoid a
initial resync. But drbdmanage now uses exactly this method, so it
is no longer uncommon.

Use the BLKZEROOUT ioctl to do it as efficient as possible on Linux
kernels 3.6 or later. On older kernels we use pwrite.