introduce new "disable-write-same" disk-option
Some backends pretend to support write-same,
but then fail the actual requests.

DRBD would then treat it as backend IO error, and detach.

You can tell the system to disable write-same on such devices:
  echo 0 | tee /sys/block/*/device/scsi_disk/*/max_write_same_blocks

But you'd need to detach and re-attach the whole stack of virtual disks
between that backend and DRBD to propagate that change.

With this new disk-option, you can disable write-same on DRBD
even though the backend claims support. Or toggle at runtime.