drbdadm adjust: Optimize the single resource adjust case
If you have 1000 resources in your kernel, and do an adjust on one
of them, then drbdadm adjust did a show of all of them and parsed
all of them from the running kernel.

-> Leads to a runtime of about 0.6 seconds.

If we parse only the single one we have a runtime in the area of 0.02 seconds.
Keep it efficinet for the "ALL" case where it should invoke drbdsetup
show only a single time.