drbdadm,v9,proxy: fix segfault in proxy-up
So far we got a segfault in proxy scenarios when we had 3 or more nodes.
This happened for example in a 3 node setup when there was a proxy
connection between A and B, A and C and B and C, and the command was
executed on node A. In that case A also tried to setup the proxy between
B and C because the code iterated over all connections and then ran into
a NULL ptr deref. This skips connections if it does not contain a valid
proxy configuration for the executing host.