v9: introduce intentional diskless
So far a diskless client did not know that it was a diskless client.
Therefore it was always presented as "Diskless" (in red). This commit
introduces this information and exposes it via status, show, events2.

status: On newer kernels intentional diskless clients are shown in
green. A new field "client" is introduces and shown with "--verbose".

show: Intentionally diskless clients are shown with "disk none".

events2: The output now also contains the new "client" field.

For parsable commands (status --json, events2), the new information is
always show. If the kernel is too old, it *always* shows the "client"
field as "unknown". In some scenarios we could decide to leave that
information out or use additional information (e.g., is the node even
diskless or not). I deliberately decided against that. It increases
complexity (if you want that new, additional information, use a new
kernel where that information is set correctly). Additionally, that way
it is easier for parsers. If you read an "unknown", you know that your
kernel is too old. If not, you know that the information ("yes"/"no")
is accurate and your kernel new enough.