v9: crm-fence-peer.9.sh: make Pacemaker CIB based fencing "multi-peer aware"
This was already a complex script that tried to be "smart" in a lot of
cases, and adapting it for DRBD 9 did only add to that complexity.

That's why I chose to add a new script intended for new DRBD 9 + Pacemaker
installations, rather than replace the existing "proven" script, even
though the new one is supposed to behave mostly "identically" to the old
one for the DRBD 8.4 respectively two-node case.