Pass md_index information to detect_md
detect_md() assumes that the metadata index format is
DRBD_MD_INDEX_FLEX_INT, but in fact it can also be
DRBD_MD_INDEX_FLEX_EXT.

[lge: Actually, it could even be old-style indexed fixed size meta data,
but that can be fixed an other day, or worked around with dump/restore.

Thanks for noticing and fixing a regression when trying to detect and
"convert" external meta data.  We introduced said regression likely in
late 2011, even before 8.4 came out.]

Signed-off-by: Kristoffer Grönlund <krig@koru.se>
Signed-off-by: Lars Ellenberg <lars.ellenberg@linbit.com>