drbd udev: fix inconsistent inheritance of implicit volumes
The "implicit-ness" of volume definitions on the resource level
was not inherited by the host sections.

That confused the generation of udev symlinks to sometimes
(but not always) inconsistently use "/0" in udev symlinks
even for implicit volumes.

Also, such volume definitions would forget to print the "disk"
definition in a non-verbose dump.

This patch fixes that, and fixes the udev symlinks so implicit volumes
would behave as intended and never add the "/0".

For those that "accidentally" had these unintended symlinks, or want
them for any reason, it adds a new global option udev-always-use-vnr,
which is disabled by default, but enabled in the example global_common.conf.

   udev-always-use-vnr

       When udev asks drbdadm for a list of device related symlinks,
       drbdadm would suggest symlinks with differing naming
       conventions, depending on whether the resource has explicit
       volume VNR { } definitions, or only one single volume with
       the implicit volume number 0:

	   # implicit single volume without "volume 0 {}" block
	   DEVICE=drbd<minor>
	   SYMLINK_BY_RES=drbd/by-res/<resource-name>
	   SYMLINK_BY_DISK=drbd/by-disk/<backing-disk-name>

	   # explicit volume definition: volume VNR { }
	   DEVICE=drbd<minor>
	   SYMLINK_BY_RES=drbd/by-res/<resource-name>/VNR
	   SYMLINK_BY_DISK=drbd/by-disk/<backing-disk-name>

       If you define this parameter in the global section, drbdadm
       will always add the .../VNR part, and will not care for
       whether the volume definition was implicit or explicit.

       For legacy backward compatibility, this is off by default,
       but we do recommend to enable it.