adjust: fix deleting unrelated peer if host was given
In a multi-node setup, we could have the following situation:
- mesh between A,B,C
- on A: disconnect res:C
- on A: adjust res:C

So far adjust then deleted peer B. This was caused by
adjust_net/matching_conn. The connection between A and B had the ignore
flag set, the existing connection between A and B was therefore not
found/returned and then the peer B was deleted.

Keep the information that it is a connection involving the host for such
connections (the "me" flag). Then tell matching_conn that it should also
return such connections and therefore not delete the peer.

Another possibility was setting the peer information also for already
ignored connections and use that information later (i.e., basically the
same, namely finding out if it is a connection involving the host). The
ignore flag is too overloaded.

If a peer gets deleted from the res file (e.g., B) and we call adjust
for C, the vanished peer B still gets deleted. IMO that is valid.