Changed the '0', "Pg0" hotkey to '1', "Pg1"
The hotkey for returning to the first page was changed
from hotkey '0', label "Pg0" to hotkey '1', label "Pg1".
Although it is technically correct that the hotkey resets
the page index to 0, as DrbdMon uses page numbers
0..65535 internally, the page numbers displayed to the
user are in the range 1..65536, so from the user's
point of view, page 0 does not exist, and the hotkey
resets the view to page 1.