Prevent write retry deadlocks just in case someone runs DrbdMon at realtime priority
The CompactDisplay retries failing or incomplete write() to the output stream
(commonly stdout). Since the cause for failing write() calls may be that whichever
process is supposed to receive the data has not yet read the data sent by DrbdMon,
DrbdMon will normally yield execution to other processes before retrying a write() call
using sched_yield().

However, in the case that someone runs DrbdMon at realtime priority, yielding execution
may have no effect if DrbdMon is the highest-priority task in the system, because it will
immediately become runnable again.
To avoid a deadlock in this case, DrbdMon will switch to suspending itself for a short
period of time (currently 20 ms) if yielding repeatedly did not lead to any progress, so
that a lower priority process that receives DrbdMon's data has a chance to run and read
the data.