fix major()/minor() use on recent glibc
Modern glibc started to spill the following warnings:

warning: In the GNU C Library, "minor" is defined by <sys/sysmacros.h>.
For historical compatibility, it is currently defined by <sys/types.h>
as well, but we plan to remove this soon. To use "minor", include
<sys/sysmacros.h> directly. If you did not intend to use a
system-defined macro "minor", you should undefine it after including
<sys/types.h>.

We had that 3 times in different tools, adding it to shared_tool.h was
the easiest to fix all of them. RHEL 5, the oldest distribution we still
support, has that include file.