drbdadm,v8.4: allow "create-md --force"
So far it was neccessary to call that command as:
"drbdadm -- --force create-md foo"

This differes from v9, where
"drbdadm create-md --force foo" is allowed. This difference makes it
unneccessary hard for users as well as tools like godrbdutils/drbdtop.

"create-md" is a strange command in that regard, that options are setup
options, but unlike real setup commands a drbdsetup_ctx does not really
make sense.

Solve that by being not strict in the parser if the command is
"create-md". In the end it does not make a difference, because invalid
options make drbdmeta then fail anyways.

This allows the old and the new form.