drbdmeta compat: fixup fallback path for BLKZEROOUT for older kernels
Based on a suggestion on drbd-dev by
 From: Artem Pogartsev <artem.pogartsev@activecloud.com>
 Date: Thu, 17 Aug 2017 04:57:12 +0300
 To: drbd-dev@lists.linbit.com
 Subject: [PATCH] drbdmeta: Fix drbdmeta to support old kernels
 Message-ID: <a1c1b596-8259-a2cf-3b1b-6461761bbca4@activecloud.com>

Though I chose to always execute the fallback code
if the ioctl BLKZEROOUT fails, regardless of errno.

As a side note,
just because the define BLKZEROOUT is not in the (userland) headers
(yet) does not necessarily mean the kernel does not know it.