drbdsetup,v9: drbdsetup events2 --now --poll
"--poll" is a new argument that is completely ignored if "--now" is
not given. If both are specified, the output of "--now" is printed, and
then the commands keeps reading single character commands. For now this
can be 'n' to re-trigger a "--now"-run. All other characters break out
of the loop, newlines are ignored.

This is usefull for commands that spawn "drbdsetup events2 --now
--statistics" very often (i.e., drbdtop). Now these tools can spawn
"drbdsetup" once and re-trigger output when they feel like it.