drbdadm,v9: fix connection-mesh and proxy combination
So far a "proxy-up" produced the following error message:
'r0' not defined in your config (for this host).

Actually proxy-up handles this already fine if the connection-mesh is
not given. I attempted to fix that by ignoring connection-mesh expansion
for hosts where a proxy is involved if there is already an "implicit
connection" generated for the proxy. To get access to the information -
and skip these hosts - in the code that expands the connection mesh it
would have been necessary to reshuffle implicit connection creation and
connection mesh expansion. This broke other use cases (needs more than
two nodes). I gave up. Error out if connection meshes are used in
combination with proxy commands. Usually you do not need the
connection-mesh in such a case. But if, you have to specify the explicit
connections yourself.

resource r0 {
   protocol A;
   device    /dev/drbd0;
   disk      /dev/testlv;
   meta-disk internal;
   proxy {
      memlimit 100M;
   }
   on rckdeba {
      address   127.0.0.1:7997;
      proxy on rckdeba {
	 inside   127.0.0.1:7998;
	 outside   172.16.7.97:7999;
      }
      node-id   0;
   }
   on rckdebb {
      address   127.0.0.1:7997;
      proxy on rckdebb {
	 inside   127.0.0.1:7998;
	 outside
	    172.16.7.98:7999;
      }
      node-id   1;
   }
   connection-mesh {
      hosts rckdeba rckdebb;
   }
}