v9,adjust,proxy: Fix adjust for proxy when proxy is on a dedicated machine
The way the logic from drbd-8.4 was extended did not took into account
that on different connections the proxy could be dedicated or local
on the same machines where drbd runs.

Therefore the can_do_proxy variable needs to be per path. Inverted
it, because these bits are initialized to 0.

In addition check if the proxy is configured to run on this host in
adjust_net. (In the 8.4 world that was implicit by the can_do_proxy
variable)