drbdadm,v9: proxy-up was broken when some of the connections have a proxy
If at least one connection is with proxy and at least one without
proxy the proxy-up command gave an error and aborted when it found
the connection without proxy.

Fixed by not aborting and faking success as long the user did
not asked explicitly for that connections. (Only fake success
if we iterate over all connections of a resource)