drbdadm,v9: Fixing a couple of issues in adjust of connections in the proxy
Consider the case that the connection is not configured et all in
the proxy. Recognise that by the parse error AND the exit code of
drbd-proxy-ctl -s "show-proxysettings <name>".

When parsing the settings from the running proxy actually set the
pointer to the new data-structure!

Now it works for my limited number of test cases.