drbdmon: Add nullptr check for potential future use
DrbdResource::remove_connection() does not check whether the
node to remove was actually found, however it is only used by
a function that has already determined that the node exists.
A nullptr check was added for potential future use of the
same method by a caller that did not verify whether or not
the specified connection name is actually mapped in the
ConnectionsMap.