drbdadm v84: fixup for disable-write-same
This is a fixup for the 8.4 part of
a3d20ef49c7a introduce new "disable-write-same" disk-option

Using "diable-write-same" on DRBD 8.4

I failed to add it to the scanner keyword list for 8.4.
Apparently putting the setting into the config file
was never tested with DRBD 8.4, but only with 9,
where modifying the scanner is no longer necessary.

"live toggling" via "drbdadm/drbdsetup disk-options" works with 8.4,
but breaks "drbdadm adjust".

Fixed now.

Reported-by: Tomoaki Sato <tsato@valinux.co.jp>