drbdsetup,v9: expose "client" if !isatty
So far if you did a "drbsetup status | less", the color information if
the disk is intentionally diskless or not was lost and no additional
information was provided. --verbose sometimes is a bit too verbose if
you want to quickly check many resources. Now a textual information is
shown via the additional "client" field (as it would be in --verbose).
So this does not clutter the default where you have colors, but provides
that valuable information if you don't.

An alternative to keep the color information (thanks Lars) is:
drbdsetup status --color=always | less -R

I still find it valuable as the quicker/normal shortcut, and overall it
removes more lines than it adds...