drbdsetup,v9: make json output a bit nicer to parse
With the new resync/verify statistics sometimes parts of them make
sense, sometimes they don't and are not even printed. So you would have
to check if keys exist in the json output. If you throw it at the json
parser on golang, you would potentially end up with default values (or
you would have to initialize the struct to a know "invalid" value). All
in all to me it feels nicer to have additional fields that are always
there that state a) that you can expect the keys, b) that it contains
meaningful data.

While at it convert all fields that intend to print propper booleans
("true"/"false") to use a helper function. This fixes the "quorum"
output, which was intended as boolean json type, put printed as string.
Still we have a bad mix of "yes"/"no" with boolean types. For now keep
as is, maybe "break" (but fix) with next release.