build: move defines to configure stage
grepping for enums and then setting defines (twice) is ugly. Move that
to the configure stage. There already was a header and a compile check.
I don't know why this other hack was introduced at all. As libgenl.c
includes the config.h (one of previous commits), we can now set these
defines directly in config.h

The only user visible change would be a v83 only build, which we really
do not care about.