Added optional display update frequency limiting
When DrbdMon receives event lines from drbdsetup, a display
update will only be executed immediately if the interval
between the display updates is at least as long as the
update frequency allows. Otherwise, a timer is set to
update the display according to the selected display
update frequency if no further event lines arrive.