doc,84: resize --size example
Strange enough in the last two days two people managed to get that wrong
(drbdadm resize --size=X ...). So give an example and mention that all
of the options are backend options.

As a side not: v9 gets that "right", but that is just how backend
options in v84 work, so fix that one in the documentation.