drbdadm,v9,init: fix help in adm, avoid output in init
It looks a bit like there could be a revival of the init script (which
is also used in the service file). For example we use it in Proxmox
setups where the LINSTOR controller is executed in a VM on DRBD storage.

There one of the nasty outputs in journalctl is the failed call to
"sh-b-pri". While ignored, the output ends in the log, which it should
for 8.4.

Fix that by checking if the command exists (i.e., by calling "help" on
it). For that to work I fixed the return codes of "help" (or "-h") for
unknown commands. So far "drbdadm doesnotexist -h" returned exit code 0.

On 8.4 that command exists anyways and drbdadm would return 0 anyways,
but on 9.0 handle it correctly and use that behavior in the init
script.