drbdsetup,v9,show,json: simplify print_paths
Here the difference between LINBIT/DRBD output and json output is
sufficiently small, so unify this function.

Introduce 2 macros that produce indented output depending on if json is
on or not.

Unfortunately the other functions that do this "_json" split are no
candidates, they differ in too many details. Better have a bit more code
duplication than a horrible if-maze.