drbd-overview: remove it (most of it)
It has been deprecated for a while and we have better tools (drbdmon and
drbdtop). I don't have to impression that people really need it.

Especially, I don't like that this pulls in Perl. In a "container
world"/small VM base-image world you don't want to pull in ~50MB of
dependencies for a script that basically nobody uses.

Remove it from packaging, but still keep it in git. If somebody really
wants it, it can be downloaded from GH/maintained in a separate project.