usage-count: only early exit if no tty and ask
So far we always early returned when there was no tty. Don't be that
strict and only early return when the usage count is set to 'ask'. This
allows us to count for example LINSTOR installations that call drbdadm,
which then has no tty for obvious reasons.

In my test this did not break any use cases. Autocompletion did not get
noisy and 'yes' and 'no' are well defined anyways. FWIW, Debian ships
with 'no', ELRepo ships with the default of 'yes'.

While at it, some small cleanup. Remove 'send'. It had a weird mix of
early return and single exit semantic.

We also checked for an environment variable that was historically set by
"init", to also early return. We don't want to export that in the
script. Admins might execute it, and there it would be perfectly fine to
ask a question. We never want systemd to stop here, so just set this
magic variable in the service file.