drbd.ocf: don't add --stacked for drbd 9
For drbd kernel module version 9,
starting with drbd utils version 9.5.0,
we don't need to add "--stacked" anymore,
respectively we ignore its presence/absence,
at least if explicitly naming the resources to act on.

Because of that, the logic in the DRBD ocf resource agent
would now always add --stacked.
Reverse the logic to first test if non-stacked works.