drbdadm,v9: fix resync-after
So far this was pretty broken for 9. It for example failed with:
resync-after contains 'FirstSyncer/0',
   which is ambiguous, since it contains 2 volumes

The idea is to stip the volume part at one location, but also check the
volume information in the second hunk.

This was tested with a resource that contains 2 volumes. The
resync-after was placed globally, as well as into one of the both volume
sections. The resync-after target was once a resource with one volume
and once one with 2 volumes. "res", and "res/volid" combinations have
been tested as well.