parser,v9: require-drbd-module-version
This introduces require-drbd-module-version-{eq,ne,gt,ge,lt,lq}, which
can be set on the global level of resource files (e.g., like
"resource"). The key is followed by a DRBD kernel module version in the
format of "major.minor.patch". For example if one requires a DRBD module
version strictly greater than 9.0.16, one can add this to a res file:

require-drbd-module-version-gt 9.0.16;

This has two purposes:
- obviously it allows us/LINSTOR to specify dependencies
- We have a lot of people that try to run LINSTOR and fail at the most
basic level, namely loading a DRBD9 module. This is a unknown token for
8.4. drbdadm parsers should give a sufficient hint what went wrong.