drbdadm,v9: disk-options should enumerate the config file settings
The disk-options subcommand, if used from the command line and not from
some internal indirection helper, was ignoring the settings from the
config file, only passing on settings listed directly on the command line.

That is not what we mean or expect, and inconsistent with the behaviour
of our other "*-options" subcommands.

Fix: do not skip the make_options() call.