drbd.ocf: new parameter connect_only_after_promote
This may be useful for "stacked" setups without proper fencing on the
lower layer (which we obviously do not recommend), to avoid some of the
ugly side effects that may arise after resolving a split-brain on the
lower layer.

Keep this DRBD instance disconnected until it is promoted.
After promotion we issue an additional "adjust",
which is supposed to initiate the connection attempts.

This causes a new data generation identifier ("current uuid")
to be generated after the failover of a "healthy" DRBD instance.