build,man: easier build integration
This introduces a new configure variable (--with-prebuiltman), which can
be set to skip (re)generation of man pages. This is helpful when one
generates packages from a release tarball. These release tarballs
already contain the man pages, we should not regenerate them. With that
in place, we can drop the build dependency on po4a.

Do not use hard-coded paths for po4a-{translate,gettextize}.

Split man page generation and pot generation. This also is nasty in the
build system, and not needed. People needing that, have to know/look for
the "pot" target.

To build packages:
- Always generate a propper release tarball and start from that one. Use
the usual "make .filelist && make rpmprep && make rpm" or a
"debuild/dpkg-buildpackage".
- If for whatever reason you want to generate a package from a git
checkout (you should NOT), you can: run the above rpm commands, or set
DEB_BUILD_OPTIONS to include noprebuiltman, as in:
DEB_BUILD_OPTIONS=noprebuiltman debuild -us -uc -i -b