tests: add regression tests
This adds an initial version of a simple regression test suite. It
executes drbdadm commands and checks if the the dry-run output matches
what is expected.

This uses "clitest". On apt based systems, you should be able to "apt
install clitest". On "enterprise distributions", you might need to fetch
the script from: https://github.com/aureliojargas/clitest

Tests can be executed via:
$ make test
or for example via
$ cd tests
$ ./run.sh ./v9/*.test

I'm pretty sure there is room for cleaning up run.sh, but it is a start.