drbdadm: Generalize paths for in stacked resources
In the connection section of a stacked resource one can reference to
host sections of the lower resources. This is nothing new.

With that it is possible to create pathes from lower hosts that are
not local to this machine. Consider that by introducing an ignore
flag on the path.

consider this config excerpt:

	stacked-on-top-of site1 {
		node-id 0;
	}
	stacked-on-top-of site2 {
		node-id 1;
	}
	stacked-on-top-of site3 {
		node-id 2;
	}

	connection { # site1 - site2
		path {
			host alpha address 192.168.23.21:7100;
			host charlie address 192.168.24.21:7100;
		}
		path {
			host bravo address 192.168.23.22:7100;
			host delta address 192.168.24.22:7100;
		}
		path {
			host alpha address 192.168.23.21:7100;
			host delta address 192.168.24.22:7100;
		}
		path {
			host bravo address 192.168.23.22:7100;
			host charlie address 192.168.24.21:7100;
		}
	}

alpha and bravo are hosts of resource site1. chalie and delta are
hosts of resource site2. Now, when bringing up the paths on
alpha the paths from bravo need to be ingroned.