docker: add dockerpath target for compat
This target is used in all the projects that build for the "docker
distribution" in lbbuild. It loops over the output and pushes these
images (to drbd.io).

So far this image is only useful as base for linstor-satellite, so no
need to push it to drbd.io.

Still, provide the target. We can switch fast, and we do not need to
special case it in lbbuild.