docker: base images on centos
While this is ways clumsier, we need to base our image on a RHEL product in
order to get certified by OpenShift. This prepares that switch.

Note that we use centos as builder, because the UBIs only have a very
limited set of packages. We build with Centos and use the pkgs in the
UBI container.

It does not make sense to push this image to drbd.io. But we need it
somewhere, as it is the base image for the linstor-satellite. The UBI
based satellite will pull from quay.io. I already pushed a version of
drbd-utils there. So this should be a smoth transition.