drbdmeta: allow huge volumes for DRBD9
We have a limit of 1 PiB for a DRBD volume. And before DRBD9 we only had
one peer. bm_bytes() does the calculation including peers, but when we
check if the bitmap would get too large, we only expect one peer.

IMO the check there is to limit the volume size to 1 PiB, but it does
not try to limit the overall MD size. Therefore, the fix is simple, just
multiply the max possible size with the number of peers.

For DRBD < 9 this is a factor of 1, so there should not be any
regressions.