drbd.ocf: new parameter fail_promote_early_if_peer_primary
The "longdesc" of the meta-data says:

To avoid a useless retry loop during promotion attempts in case of cluster
split-brain (Pacemaker communication loss) while DRBD is still connected to a
Primary, you can chose to give up after the first try if this situation is detected.

If a Primary "vanishes", TCP may not immediately detect this, and an idle DRBD
may take some time until it does in-DRBD-protocol "pings". Pacemaker may well
detect Primary loss earlier than DRBD, and try to promote while DRBD thinks it
can still see a Primary. Which means, in general, trying to promote at least
once is necessary, as that implies an in-DRBD-protocol "peer alive" check.

But if that does not succeed, re-trying until we hit the operation timeout may
not be desired, so you can disable it.