drbdsetup,json: fix output if there are no volumes.
So far we did not print a ',' if there are no volumes in a resource
after the node-id. This can be fixed in two ways:

- don't print the volume section at all if there are no volumes. This
was proposed on drbd-user. This potentially breaks if there are other
sections.

- put a ',' after the last section (node-id), and then always print the
volume section. If there is no volume, just print an empty '"volumes":
[]', which is fine, and which I actually prefer. So one can always
query the keyword, but the array might be empty.

Reported-by: Anton Tykhyy <atykhyy@gmail.com>