crm-fence-peer: detect pacemaker 2 "promotable clones"
The sed-script to guess the pacemaker xml-id to be used in location
constraints from the DRBD resource name expected '<master' tags, but
with new pacemaker 2 xml, these are '<clone's, with "promotable=true"
set in the meta attributes.

Use the crm_feature_set version from the top level <cib/> tag
to decide between the two.