scripts: Blacklist drbd for multipath / multipathd
I have heard people saying that multipath kept their drbd devices open
(read only) and therefore made it impossible to promote it on another node.
That was on Ubuntu focal.

I witnessed the same on a RHEL7 system.

Add a multipath config snippet that blacklists all drbd device nodes for
multipath and multipathd.