drbdsetup,events2: make may_promote more strict
As it was the may_promote shows "yes" if the local disk is D_OUTDATED.

The disk status can display D_OUTDATED due to two reasons in DRBD:
 1) There is a primary node that is not a neighbor to this node. I.e. it
    is connected through at least one intermediate node.
 2) The node got disconnected from the partition with the primary and
    a fencing handler got executed changed the disk state from
    D_UP_TO_DATE to D_OUTDATED.
    A characteristic of the second case is that there is no D_UP_TO_DATE peer.

I was tempted to express it as:
	info.may_promote = !have_primary(resource) && disk_state != D_OUTDATED;

But that does not work as the local disk state might fall to
D_INCONSISTENT and due to a resync and goes back to D_OUTDATED as long
a primary node is present (far away) in the partition.

Now, the "may_promote" is a bit more conservative than necessary.
I.e. it will say "may_promote:no" for a node that is resync target.
But that is no problem for the purpose of the "may_promote" field,
because in that case the sync-source node will show "may_romote:yes".
It is sufficient if it says "yes" on one node of the surviving partition.

The purpose I refer here to is that a partition can recognize when the
former primary fails and it is guaranteed that this former primary has lost
quorum. For that, a "false negative" on a node is not a big issue, while
a "false positive" would cause troubles. (Troubles in the form of trying
to start the service, to promote the node to primary, etc...)