drbdsetup,events2: restore the ordering of lines
Before the big change the lines of a state change always had the order:

resource changes
connection changes
device changes
peer device changes

with the new implementation that changed to:

resource changes
device changes
connection changes
peer device changes

With that patch, we are going back to the old ordering. The reason is
mainly making it compatible with the assumptions of the test-suite. None
of these ordering is better or more correct. Just staying as it was
before makes live easier.