crm-fence-peer.9.sh: fix Pacemaker compat
Trying to parse the crm_mon output intended for "human consumption" was
a bad idea, unsurprisingly.

Parsing the `crm_mon --as-xml` variant should turn out more stable.
This fixes additional regressions in more scenarios that should be
covered by this "crm-fence-peer" handler, where it needs to check if the
node that executes the handler is already rejected by an existing
constraint with the expected constraint id.

Now works again with DRBD 9 and "all" (relevant) versions of
pacemaker, where "relevant" is 1.1.x >= 1.1.15,
and the current 2.0 series.