drbdsetup,events2: handle mixed initial state and multicast events
We join DRBD's multicast group before requesting the initial state.
This is necessary to avoid missing any state changes, but means that
normal events can be mixed in with the initial state. In addition, the
initial state is a snapshot of the state at some point after we request
it, which is sent without synchronization with the multicast events.

This means that it is possible for various counterintuitive scenarios to
occur. Events other than 'exists' may be received before the initial
state is done. These events may or may not apply to an older version of
the state than the one in the initial state. Also, we may receive
'create' and 'exists' for the same object. Or similarly, 'destroy'
events for objects for which we never receive 'exists' events. In this
case, there may even be 'change' events before the 'destroy' events. It
is even possible to receive the complete initial state including the
intial state done message, and then receive events for changes that are
already represented in that initial state.

The way that events2 is implemented was significantly changed in commit
09c5ee367513 drbdsetup,events2: keep track of tree of objects

Prior to this, events were simply formatted and printed as they were
received. Now they are applied to the tree of objects and then the
differences are printed.

With the new implementation, when these counterintuitive scenarios
occurred, this led to errors of the form:
resource name:res0 not found
or
resource name:res0 already exists

Fix this by:
1) Buffering the multicast events until the initial state is done. Now
the 'exists' events are always printed before any others. This makes it
easier to use the events2 output.
2) Being less strict in our validation. We essentially drop creation
events that refer to existing objects, as well as change or destruction
events that refer to non-existent objects.

This gives us an eventually consistent view of the object states. There
are still corner cases that could lead to events being shown which
correspond to a state which never actually existed, but this is the best
we can do with the API currently available to us.