initial "drbd-attr" pacemaker resource agent
This parses the "drbdsetup events2" stream for the "promotion_score",
and injects this as transient node attribute into Pacemaker.

We can now add location constraints referencing these node attributes.

Intended usage is a setup with LINSTOR and Pacemaker, where we don't
want to know which nodes have access to which DRBD resources,
and also don't want to handle configuration / promotion of DRBD
in Pacemaker, but use auto-promote.

To avoid spurious auto-promote failures, we still need to make Pacemaker
aware of the per-node "ready-ness" (both during operation, and during
"boot up" of these resources).