drbdsetup,v9: fix a Wmaybe-uninitialized
without it, we get a warning on recent gcc:

drbdsetup.c:3632:10: warning: ‘tail’ may be used uninitialized in this
function [-Wmaybe-uninitialized]
 3632 |   **tail = p;
      |   ~~~~~~~^~~
drbdsetup.c:3879:43: note: ‘tail’ was declared here
 3879 |  struct peer_devices_list *peer_device, **tail;
      |                                           ^~~~

This must be pretty esoteric as I never saw any issues with it. Still,
the compiler is actually right. Initialize the pointer and check in the
callee.

While the compiler did not complain at the other remember_* functions,
add such a check there as well.