drbdsetup,v9: respect the 'stable' flag for may_promote
This fixes cases where the primary node is separated from us by 2 steps.
In this situation we may have peers with up-to-date data. We see no
primary directly, however we must not report may_promote:yes. Solve this
by only allowing may_promote:yes when we have access to _stable_
up-to-date data.