doc,v9: improve connection section in drbd.conf
This does 2 things:
- mark the net section as optional
- remove the indenting for the volume section as it is on the same level
  as for example the net section, but not part of it.

While at it also mark net in connection-mesh as optional and clarify the
path/host situation a bit.

All in all this still is not bullet proof, as most of the other sections
unrelated to connection would have to be marked as optional. Even more
difficult to express, some "things" have to be set on some level, but it
is up to the user where they are set. This simple representation is not
expressive enough to do that.