drbdadm,v9: allow set-gi in single node envs
So far set-gi simply was not executed at all. LINSTOR - which directly
calls drbdmeta - uses its own node-id in such a case. This mimics this
behavior.

LINSTOR has it's own data base, it is easy for it to construct all the
drbdmeta parameters. In a more general library (i.e., libdrbd-perl),
that has no (and should not have) any knowledge about res file
interpretation, this gets a lot nastier with tons of sh-* calls.

If we assume starting with a single node and then extending it and
skipping the initial sync (why set-gi is used in the first place), then
this is a bug that should be fixed.

resource "rck" {
    volume 0 {
        disk /dev/local/rck;
        meta-disk internal;
        device minor 23;
    }
    on aperol {
        address ipv4 10.43.70.1:2342;
        node-id 2;
    }
    # on aperol2 {
    #     address ipv4 10.43.70.2:2342;
    #     node-id 1;
    # }
    connection-mesh {
        hosts aperol;
    }
}

results in:
drbdmeta 23 v09 /dev/local/rck internal --node-id=2 ...
instead of ignoring it.

enabling aperol2 results in:
drbdmeta 23 v09 /dev/local/rck internal --node-id=1 ...
which is the old behavior.